+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.32.1
+### Bugfix
+* leading spaces with echo + keepWords [#754](https://github.com/jcubic/jquery.terminal/issues/754)
+* fix loading hidden terminal
+
 ## 2.32.0
 ### Features
 * add insert typing animation